requirements.txt
flask==2.3.3
werkzeug==2.3.7
python-dotenv==0.19.0
deepgram-sdk==2.11.0
python-multipart==0.0.5 